Views Refactor
- Added views directory and split IndexView and OrganizationView into separate files
- Updated organization and score seed files
