+++ crank/views/organization.py
@@ -1,0 +1,13 @@
+from django.views import generic
+
+# from crank.models.score import Score
+from crank.models.organization import Organization
+
+
+class OrganizationView(generic.DetailView):
+    model = Organization
+    template_name = "crank/organization.html"
+
+    def get_queryset(self):
+        """make sure we can't see inactive organizations."""
+        return Organization.objects.filter(status=1)
